Conflicts:
	RootSystems/Include/Components.h

+++ RootEngine/Source/EffectImporter.cpp
@@ -78,7 +78,7 @@ namespace RootEngine
 				{
 					std::vector<std::string> varyingsVector;
 					const YAML::Node& varyings = programs[i]["feedback"];
-					for(int k = 0; k < varyings.size(); k++)
+					for(size_t k = 0; k < varyings.size(); k++)
 					{
 						std::string varying;
 						varyings[k]["name"] >> varying;
